# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pterodactyl Backup Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface BotConfig {
  panelUrl: string;
  username: string;
  password: string;
  serverId: string;
  backupIntervalMinutes?: number;
}

export interface BotStatus {
  running: boolean;
  configured: boolean;
  /** @nullable */
  lastBackupAt: string | null;
  backupCount: number;
  /** @nullable */
  currentAction: string | null;
  /** @nullable */
  uptime: number | null;
}

export interface TelegramInitInput {
  apiId: number;
  apiHash: string;
  phoneNumber: string;
}

export interface TelegramOtpInput {
  code: string;
}

export interface TelegramStatus {
  connected: boolean;
  authenticated: boolean;
  awaitingOtp: boolean;
  /** @nullable */
  phoneNumber: string | null;
}