# lib/api-zod/src/generated/types/botConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pterodactyl Backup Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfig {
  panelUrl: string;
  username: string;
  password: string;
  serverId: string;
  backupIntervalMinutes?: number;
}