# lib/api-zod/src/generated/types/botStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pterodactyl Backup Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BotStatus {
  running: boolean;
  configured: boolean;
  /** @nullable */
  lastBackupAt: string | null;
  backupCount: number;
  /** @nullable */
  currentAction: string | null;
  /** @nullable */
  uptime: number | null;
}