# lib/api-zod/src/generated/types/telegramInitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pterodactyl Backup Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramInitInput {
  apiId: number;
  apiHash: string;
  phoneNumber: string;
}