# lib/api-zod/src/generated/types/telegramStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pterodactyl Backup Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramStatus {
  connected: boolean;
  authenticated: boolean;
  awaitingOtp: boolean;
  /** @nullable */
  phoneNumber: string | null;
}